backend/requirements.txt
flask
flask-cors
numpy
pillow
torch
torchvision
scikit-learn
joblib
huggingface_hub
opencv-python